refactor(logic): split connectivity_resolver into part-of concern files (Refs #3290)
Phase-0 prerequisite for the colonizethis_logic package split: decompose the
711-line world/connectivity_resolver.dart into cohesive part-of files, each
under 400 lines, with zero behavior change (same library scope, identical
public API and library-private symbols).

- connectivity_metrics.dart: hot-path metrics counters + test hook
- connectivity_tile_helpers.dart: pure tile/transport grid helpers
- connectivity_propagation.dart: BFS propagation, sea-port wiring, town-rule closure
- connectivity_resolver.dart: public entry points + per-player province cache

Verified: dart analyze clean; all 39 connectivity tests pass; logic dead-file
checker passes (new part files recognized).
